Expand description of denial of existence.
Also change MUST use DANE when TLSA records are present to SHOULD
use DANE when TLSA records are present AND "usable".  With the
protocol subject to downgrade attacks, the MUST does not afford any
downgrade resistance.  The client may do as it pleases and the
server is none the wiser.